refactor: icon components #67
